+++ casiopea-constel.alias.php
@@ -10,11 +10,13 @@ $specialPageAliases = [];
 /** English */
 $specialPageAliases['en'] = [
 	'Constellation' => [ 'Constellation' ],
+	'ConstellationSvg' => [ 'ConstellationSvg' ],
 	'MyConstel' => [ 'MyConstel' ],
 ];
 
 /** Spanish (español) */
 $specialPageAliases['es'] = [
 	'Constellation' => [ 'Constelación', 'Constelacion' ],
+	'ConstellationSvg' => [ 'ConstelaciónSVG' ],
 	'MyConstel' => [ 'MiConstel' ],
 ];